Create a custom hook as a Store which will be responsible for global state management
GlobalState, listeners and actions are outside of the hook and they will be the same for every component that uses this file. They are block scoped and not globally available variables.

We are using useState here cause  it will provide us this mechanism that will re render the component that uses this hook every time the useState is changed.
This state will be the globalState. GlobalState variable is not recreated when useStore is called, it is not recreated separetely for every component that consumes the custom hook because it is defined outside of the hook. It will be created once when this file is imported somewhere. So it will share data between all files that import from it, because it is outside of the hook, whatever is inside the hook it will be exclusively to each component.

Use useEffect to register listeners when the components is updated and remove listeners when component is unmount. The listeners array will grow every time a component uses this custom hook and eventually uses the setState. We are using a closure here so setState value is captured here for that specific component that is using the custom hook and therefore will be the same when component mounts and umnounts.