Dispatch the actions
The actions is an object with keys and functions as values. The functions will return the updated globalState. So we pass the globalState to the functions as an argument "actions[actionIdentifier](globalState);". The function will return the new upated State and then we merger the global state with the newState.

After that we inform the listeners about that state update. Listener is updating the state with the new globalState and the compoent will be rerendered.