Optimize the hook store
ProductItem component rerenders all items everytime we favorite just an item. It should only rerender the item that changed. If we wrap it in React memo to prevent it from rerendering  when its props are not change it will not work, it will keep rerendering all 4 items even if we just changed the one item.
This is because each product item is using the hook store and the hook store is using the setState, so whenever the setState is being called the component that uses the hook is being rerendered. The productItem does not need to use the setState it only uses the dispatch. So by adding an argument "shouldListen" to the hook we can control if a component is using dispatch only or useState. By setting this to false in productItem then the component will not register a listener for the setState so will not be rerendered.